Improving amm position manager script and gateway http client

+++ hummingbot/core/gateway/gateway_http_client.py
@@ -830,7 +830,7 @@ class GatewayHttpClient:
         network: str,
         wallet_address: str,
         pool_address: str,
-        percentage_to_remove: float,
+        percentage_to_remove: Decimal,
         fail_silently: bool = False,
     ) -> Dict[str, Any]:
         """
